move metadata classes to their own file and update imports
